+++ src/ecc_key_local.h
@@ -59,10 +59,10 @@ struct zpc_ec_key {
 };
 
 int ec_key_clr2sec(struct zpc_ec_key *ec_key, unsigned int flags,
-			const unsigned char *pubkey, unsigned int publen,
-			const unsigned char *privkey, unsigned int privlen);
+			const unsigned char *pubkey, size_t publen,
+			const unsigned char *privkey, size_t privlen);
 int ec_key_sec2prot(struct zpc_ec_key *, enum ec_key_sec sec);
 int ec_key_check(const struct zpc_ec_key *);
 int ec_key_clr2prot(struct zpc_ec_key *ec_key, const unsigned char *privkey,
-			unsigned int privlen);
+			size_t privlen);
 #endif
